Cardiac Atlas Data

A series of shapefile data that represent cardiac anatomy in 2D.

Usage

cardiac_atlas

Format

A named list of different cardiac views, each with their underlying shapefiles. All of these shapes are a combination of line and polygon geometries. The types are the following: A

extent = map background that the shapes are built upon

chambers = individual anatomical chambers within the heart

conduction = elements of the conduction system

projections = involutions and foldings that compose critical tissue within the chamber cavities

septum = dividing structures between chambers

valves = cardiac valves (tricuspid, mitral/bicuspid, aortic, pulmonic)

vessels = venous and arterial structures

The current projections that are available under the cardiac_atlas object are:

right_atrium_rao = right anterior oblique view focused on the right atrium

Source

Created by author using the QGIS software with templates created from the "Atlas of Cardiac Anatomy" by Mori and Shivkumar (2022).

Author

Anish S. Shah, MD/MS